+++ eyes.sdk.core/src/main/java/com/applitools/eyes/visualgrid/model/EmulationBaseInfo.java
@@ -1,8 +1,11 @@
 package com.applitools.eyes.visualgrid.model;
 
+import com.applitools.eyes.RectangleSize;
+
 public abstract class EmulationBaseInfo {
 
     ScreenOrientation screenOrientation;
+    RectangleSize size = null;
 
     EmulationBaseInfo(ScreenOrientation screenOrientation) {
         this.screenOrientation = screenOrientation;